Support for Satochip applet v0.12

+++ contrib/build-wine/unsign.sh
@@ -21,16 +21,23 @@ echo "Found $(ls dist/*.exe | wc -w) files to verify."
 for mine in $(ls dist/*.exe); do
     echo "---------------"
     f=$(basename $mine)
-    echo "Downloading https://download.electrum.org/$version/$f"
-    wget -q https://download.electrum.org/$version/$f -O signed/$f
+    if test -f signed/$f; then
+        echo "Found file at signed/$f"
+    else
+        echo "Downloading https://download.electrum.org/$version/$f"
+        wget -q https://download.electrum.org/$version/$f -O signed/$f
+    fi
     out="signed/stripped/$f"
     # Remove PE signature from signed binary
     osslsigncode remove-signature -in signed/$f -out $out > /dev/null 2>&1
     chmod +x $out
     if cmp -s $out $mine; then
-	echo "Success: $f"
-	gpg --sign --armor --detach signed/$f
+        echo "Success: $f"
+        #gpg --sign --armor --detach signed/$f
     else
-	echo "Failure: $f"
+        echo "Failure: $f"
+        exit 1
     fi
 done
+
+exit 0
